@@ -1,7 +1,7 @@
 {
   "name": "vite-portfolio",
   "private": true,
-  "version": "0.0.0",
+  "version": "1.0.0",
   "type": "module",
   "scripts": {
     "dev": "vite",
@@ -10,33 +10,32 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@material-ui/core": "^4.12.4",
-    "@material-ui/icons": "^4.11.3",
-    "@react-spring/three": "^9.7.3",
-    "@react-three/drei": "^9.92.6",
-    "@react-three/fiber": "^8.15.12",
-    "@react-three/postprocessing": "^2.15.11",
-    "@types/material-ui": "^0.21.12",
-    "bootstrap": "^5.3.1",
-    "emailjs-com": "^3.2.0",
-    "react": "^18.2.0",
-    "react-bootstrap": "^2.8.0",
-    "react-dom": "^18.2.0",
-    "react-helmet-async": "^1.3.0",
-    "react-icons": "^4.10.1",
-    "react-reveal": "^1.2.2",
-    "react-router-dom": "^6.15.0",
-    "three": "^0.160.0",
-    "typewriter-effect": "^2.20.1"
+    "@mui/material": "^7.1.1",
+    "@mui/styles": "^6.4.12",
+    "@react-spring/three": "^10.0.1",
+    "@react-three/drei": "^10.2.0",
+    "@react-three/fiber": "^9.1.2",
+    "@react-three/postprocessing": "^3.0.4",
+    "bootstrap": "^5.3.6",
+    "@emailjs/browser": "^4.4.1",
+    "framer-motion": "^12.18.1",
+    "react": "^19.1.0",
+    "react-bootstrap": "^2.10.10",
+    "react-dom": "^19.1.0",
+    "react-helmet-async": "^2.0.5",
+    "react-icons": "^5.5.0",
+    "react-router-dom": "^7.6.2",
+    "three": "^0.177.0",
+    "typewriter-effect": "^2.22.0"
   },
   "devDependencies": {
-    "@types/react": "^18.2.15",
-    "@types/react-dom": "^18.2.7",
-    "@vitejs/plugin-react-swc": "^3.3.2",
-    "eslint": "^8.45.0",
-    "eslint-plugin-react": "^7.32.2",
-    "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-react-refresh": "^0.4.3",
+    "@types/react": "^19.1.8",
+    "@types/react-dom": "^19.1.6",
+    "@vitejs/plugin-react-swc": "^3.10.2",
+    "eslint": "^9.29.0",
+    "eslint-plugin-react": "^7.37.5",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.20",
     "vite": "^6.3.5"
   }
 }
